修正数据流读取字节和压缩编码整数时没有考虑长度不足的BUG

+++ Stream.h
@@ -70,7 +70,7 @@ public:
 	String ReadString();
 	//bool WriteString(const String& str);
 
-	byte	ReadByte();
+	int		ReadByte();
 	ushort	ReadUInt16();
 	uint	ReadUInt32();
 	UInt64	ReadUInt64();
